+++ spec/runSVGDomTest.js
@@ -12,7 +12,8 @@ jasmine.loadConfig({
   "spec_dir": "spec/",
   "spec_files": [
     "spec/types/*.js",
-    "spec/utils/*.js"
+    "spec/utils/*.js",
+    "spec/elements/*.js"
   ],
   "helpers": [
     "setupSVGDom.js"